Edit/delete works with scheduled

+++ src/Server/Server.java
@@ -1,6 +1,7 @@
 package Server;
 
 import ControlPanel.User;
+import ControlPanel.schedule.CalanderScheduleGUI;
 import Server.Reply.*;
 import Server.Request.*;
 
@@ -571,7 +572,7 @@ public class Server {
                         // if the user is the owner of the billboard
                         if (checkOriginOwner) {
                             // if the user has edit all billboards or create billboard permission
-                            if (temp.getEditAllBillboardsPermission() || temp.getCreateBillboardPermission()) {
+                            if (temp.getEditAllBillboardsPermission() || (temp.getCreateBillboardPermission() && !CalanderScheduleGUI.IsCurrentlyScheduled(BillBoardName))) {
                                 bb.EditBillboard(BillBoardName);
                                 EditBBReply editBBReply = new EditBBReply(temp.getSessionToken(), bb.textColour, bb.backgroundColour, bb.message,
                                         bb.image, bb.information, bb.informationColour);
@@ -627,7 +628,7 @@ public class Server {
                     // if the user is the owner of the billboard
                     if (checkOriginOwner) {
                         // if the user has edit all billboards or create billboard permission
-                        if (temp.getEditAllBillboardsPermission() || temp.getCreateBillboardPermission()) {
+                        if (temp.getEditAllBillboardsPermission() || (temp.getCreateBillboardPermission() && !CalanderScheduleGUI.IsCurrentlyScheduled(BillBoardName))) {
                             bb.DeleteBillboard(temp.getBillboardName());
                             generalReply = new GeneralReply(sessionToken,true);
                         }
